cosmetics
long line, debug warning, remove out of date TODO

+++ src/dec/webp.c
@@ -259,7 +259,7 @@ VP8StatusCode WebPParseHeaders(const uint8_t** data, uint32_t* data_size,
   }
 
   *bytes_skipped = (uint32_t)(buf - *data);
-  assert(buf - *data < MAX_CHUNK_PAYLOAD);
+  assert((uint64_t)(buf - *data) < MAX_CHUNK_PAYLOAD);
   assert(*bytes_skipped == *data_size - buf_size);
   *data = buf;
   *data_size = buf_size;
